requirements.txt
pandas>=2.0
numpy>=1.24
sqlalchemy>=2.0
pymysql>=1.1
python-dotenv>=1.0
rich>=13.0
pyyaml>=6.0
great-expectations>=0.18
matplotlib>=3.7
openpyxl>=3.1
